requirements.txt
ipywebrtc==0.4.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
multiprocess==0.70.9
pydocumentdb>=2.3.3
tqdm==4.32.2
regex==2020.2.20
azureml-sdk[automl,notebooks,contrib]==1.0.85
sklearn-crfsuite>=0.3.6
scipy>=1.0.0
numpy>=1.13.3
azureml-mlflow==1.0.85
pytorch-cpu>=1.0.0
numba>=0.38.1
ipykernel>=4.6.1
jsonlines>=1.2.0
pytorch-pretrained-bert>=0.6
requests==2.22.0
papermill==1.2.1
nltk>=3.4
tensorboardX==1.8
tensorflow-hub==0.7.0
jupyter>=1.0.0
matplotlib>=2.2.2
cached-property==1.5.1
h5py>=2.8.0
requests-oauthlib==1.2.0
tensorflow==1.15.0
python==3.6.8
seaborn>=0.9.0
Cython>=0.29.13
pyrouge>=0.1.3
azureml-dataprep==1.1.8
googledrivedownloader>=0.4
methodtools
torchtext>=0.4.0
indic-nlp-library>=0.6
azureml-widgets==1.0.85
nteract-scrapbook>=0.2.1
spacy==2.1.8
allennlp==0.8.4
seqeval>=0.0.12
dask[dataframe]==1.2.2
transformers==2.9.0
cudatoolkit=10.1
azureml-train-automl==1.0.85
scikit-learn>=0.19.0,<=0.20.3
cuda90
black>=18.6b4
-e git+https://github.com/microsoft/unilm.git@s2s-ft.v0.3#egg=s2s-ft&subdirectory=s2s-ft
pip>=19.1.1
pytorch==1.0.0
pandas>=0.24.2
py-rouge>=1.1
pytorch==1.4.0
pyemd==0.5.1
gensim>=3.7.0
pre-commit>=1.14.4
pytest>=3.6.4